requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
PyPDF2==3.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
psutil==5.9.8
numpy>=1.21.0
requests>=2.28.0
google-generativeai>=0.3.0
together>=0.2.0
pdfplumber>=0.8.0
aiofiles>=23.0.0
faiss-cpu>=1.7.4

# Additional dependencies
chromadb
pandas
uvloop
